chore(desktop): remove the retired onboarding wizard views
The live first-run path is SBOnboardingView. These SwiftUI step views have been unmounted since the Second Brain redesign; keep the shared persistence helper and GIF view that still have live callers.
