--- desktop/macos/Desktop/Sources/Onboarding/OnboardingHowDidYouHearStepView.swift
@@ -1,135 +1,0 @@
-import OmiTheme
-import SwiftUI
-
-struct OnboardingHowDidYouHearStepView: View {
-  @ObservedObject var graphViewModel: MemoryGraphViewModel
-  let stepIndex: Int
-  let totalSteps: Int
-  let onContinue: () -> Void
-  let onForceComplete: (() -> Void)?
-
-  @AppStorage("onboardingHowDidYouHearSource") private var selectedSource: String = ""
-  /// True when the step appeared with an answer already saved (a revisit).
-  /// Only the first-ever selection auto-advances; revisits use Continue so
-  /// changing your saved answer doesn't yank you forward.
-  @State private var hadSelectionOnAppear = false
-  @State private var advanceTask: Task<Void, Never>?
-
-  enum SourceGlyph: Equatable {
-    case emoji(String)
-    case youtube
-    case productHunt
-  }
-
-  static let sources: [(name: String, glyph: SourceGlyph)] = [
-    ("Social media", .emoji("📱")),
-    ("YouTube", .youtube),
-    ("Friend", .emoji("👋")),
-    ("Search engine", .emoji("🔍")),
-    ("AI chat", .emoji("✨")),
-    ("Podcast", .emoji("🎙️")),
-    ("Colleague", .emoji("🧑‍💻")),
-    ("Product Hunt", .productHunt),
-    ("Other", .emoji("💬")),
-  ]
-
-  var body: some View {
-    OnboardingStepScaffold(
-      graphViewModel: graphViewModel,
-      stepIndex: stepIndex,
-      totalSteps: totalSteps,
-      eyebrow: "Quick question",
-      title: "How did you hear\nabout Omi?",
-      description: "",
-      layoutMode: .centered,
-      onForceComplete: onForceComplete
-    ) {
-      VStack(spacing: OmiSpacing.md) {
-        FlowLayout(spacing: OmiSpacing.sm) {
-          ForEach(Self.sources, id: \.name) { source in
-            OnboardingSelectableChip(
-              title: source.name,
-              leading: AnyView(glyphView(source.glyph)),
-              isSelected: selectedSource == source.name
-            ) {
-              selectedSource = source.name
-              AnalyticsManager.shared.onboardingHowDidYouHear(source: source.name)
-              // Persist to the backend user record, not just local @AppStorage +
-              // analytics — otherwise the answer is lost on reinstall / other devices.
-              Task {
-                do {
-                  try await APIClient.shared.updateOnboardingAcquisitionSource(source.name)
-                } catch {
-                  NSLog("Onboarding: failed to persist how-did-you-hear source: \(error)")
-                }
-              }
-              // First-ever answer auto-advances; on a revisit the user changes
-              // the saved selection and moves on with the Continue button.
-              if !hadSelectionOnAppear {
-                advanceTask?.cancel()
-                advanceTask = Task {
-                  try? await Task.sleep(nanoseconds: 250_000_000)
-                  guard !Task.isCancelled else { return }
-                  onContinue()
-                }
-              }
-            }
-          }
-        }
-
-        HStack(spacing: OmiSpacing.md) {
-          OnboardingBackButton()
-
-          if hadSelectionOnAppear {
-            Button("Continue", action: onContinue)
-              .buttonStyle(InkButtonStyle(kind: .primary))
-              .keyboardShortcut(.defaultAction)
-          }
-        }
-        .padding(.top, OmiSpacing.sm)
-      }
-      .frame(maxWidth: .infinity, alignment: .center)
-      .onAppear {
-        hadSelectionOnAppear = !selectedSource.isEmpty
-      }
-      .onDisappear {
-        advanceTask?.cancel()
-      }
-    }
-  }
-
-  /// The leading mark on each chip.
-  ///
-  /// The two drawn marks keep their brand fill and their **white** knockout, and that is deliberate
-  /// rather than a missed conversion: the glyph sits on YouTube red and Product Hunt orange, not on
-  /// the glass, so `Ink.primary` here would set a near-black play triangle on saturated red. These
-  /// are the only two `.white`s in this file and neither is on the panel.
-  @ViewBuilder
-  private func glyphView(_ glyph: SourceGlyph) -> some View {
-    switch glyph {
-    case .emoji(let emoji):
-      Text(emoji)
-        .font(.system(size: 13))
-    case .youtube:
-      ZStack {
-        RoundedRectangle(cornerRadius: 3.5, style: .continuous)
-          .fill(Color(red: 1, green: 0, blue: 0))
-          .frame(width: 16, height: 12)
-        Image(systemName: "play.fill")
-          .font(.system(size: 6))
-          .foregroundColor(.white)
-      }
-    case .productHunt:
-      ZStack {
-        Circle()
-          .fill(Color(red: 0xDA / 255, green: 0x55 / 255, blue: 0x2F / 255))
-          .frame(width: 14, height: 14)
-        Text("P")
-          .font(.system(size: 9, weight: .bold, design: .rounded))
-          .foregroundColor(.white)
-      }
-    }
-  }
-}
-
-// Uses FlowLayout from AppsPage.swift
