+++ desktop/macos/Desktop/Tests/GoalGenerationServiceTests.swift
@@ -22,17 +22,11 @@ final class GoalGenerationServiceTests: XCTestCase {
       .deletingLastPathComponent()
       .deletingLastPathComponent()
       .appendingPathComponent("Sources")
-    let chat = try String(
-      contentsOf: sources.appendingPathComponent("Onboarding/OnboardingChatView.swift"),
-      encoding: .utf8
-    )
     let paged = try String(
       contentsOf: sources.appendingPathComponent("Onboarding/OnboardingPagedIntroCoordinator.swift"),
       encoding: .utf8
     )
 
-    XCTAssertTrue(chat.contains("source: \"user\""))
-    XCTAssertFalse(chat.contains("source: \"onboarding_\\(source)\""))
     XCTAssertTrue(paged.contains("source: \"user\""))
     XCTAssertFalse(paged.contains("source: \"onboarding_step_flow\""))
   }
