+++ desktop/macos/changelog/unreleased/20260901-onboarding-dead-code.json
@@ -1,0 +1,1 @@
+{"kind": "none"}
