+++ desktop/macos/e2e/flows/harness-smoke.yaml
@@ -35,10 +35,6 @@ covers:
   - desktop/macos/Desktop/Sources/OmiSupport/DesktopDevRuntimeManifest.swift
   - desktop/macos/Desktop/Sources/OmiSupport/DesktopLocalProfile.swift
   - desktop/macos/Desktop/Sources/ViewModelContainer.swift
-  - desktop/macos/Desktop/Sources/Onboarding/OnboardingChatView.swift
-  - desktop/macos/Desktop/Sources/Onboarding/OnboardingFloatingBarDemoView.swift
-  - desktop/macos/Desktop/Sources/Onboarding/OnboardingVoiceDemoView.swift
-  - desktop/macos/Desktop/Sources/Onboarding/OnboardingVoiceShortcutStepView.swift
   - desktop/macos/Desktop/Sources/Telemetry/SentryHeartbeatTelemetry.swift
   - desktop/macos/Desktop/Sources/ViewExporter.swift
   - desktop/macos/Desktop/Sources/Generated/OmiApi.generated.swift
