+++ desktop/macos/e2e/flows/onboarding-flow.yaml
@@ -8,23 +8,9 @@ app: non-prod
 covers:
   - desktop/macos/Desktop/Sources/Automation/DesktopAutomationFirstUsePopupActions.swift
   - desktop/macos/Desktop/Sources/Onboarding/SecondBrain/ThreeDoorsDemoPage.swift
-  - desktop/macos/Desktop/Sources/Onboarding/OnboardingNotificationStepView.swift
   - desktop/macos/Desktop/Sources/Onboarding/OnboardingFlow.swift
-  - desktop/macos/Desktop/Sources/Onboarding/OnboardingStepScaffold.swift
-  - desktop/macos/Desktop/Sources/Onboarding/OnboardingView.swift
+  - desktop/macos/Desktop/Sources/Onboarding/OnboardingChatPersistence.swift
   - desktop/macos/Desktop/Sources/Onboarding/OnboardingPromptSuggestions.swift
-  - desktop/macos/Desktop/Sources/Onboarding/OnboardingExportsStepView.swift
-  - desktop/macos/Desktop/Sources/Onboarding/OnboardingDataSourcesStepView.swift
-  - desktop/macos/Desktop/Sources/Onboarding/OnboardingFileScanStepView.swift
-  - desktop/macos/Desktop/Sources/Onboarding/OnboardingFloatingBarShortcutStepView.swift
-  - desktop/macos/Desktop/Sources/Onboarding/OnboardingGoalStepView.swift
-  - desktop/macos/Desktop/Sources/Onboarding/OnboardingHowDidYouHearStepView.swift
-  - desktop/macos/Desktop/Sources/Onboarding/OnboardingLanguageStepView.swift
-  - desktop/macos/Desktop/Sources/Onboarding/OnboardingNotificationStepView.swift
-  - desktop/macos/Desktop/Sources/Onboarding/OnboardingPermissionStepView.swift
-  - desktop/macos/Desktop/Sources/Onboarding/OnboardingTasksStepView.swift
-  - desktop/macos/Desktop/Sources/Onboarding/OnboardingTrustStepView.swift
-  - desktop/macos/Desktop/Sources/Onboarding/OnboardingWelcomeStepView.swift
   - desktop/macos/Desktop/Sources/PostOnboardingPromptViews.swift
   - desktop/macos/Desktop/Sources/FileIndexing/FileIndexingView.swift
   - desktop/macos/Desktop/Sources/Onboarding/OnboardingAnswerWriteGate.swift
