+++ desktop/macos/e2e/flows/onboarding-smoke.yaml
@@ -4,8 +4,8 @@ tier: manual
 description: "Manual only: reset_onboarding restarts the app. State corruption fixed in Wave 7 — run twice locally before promoting to T2."
 app: non-prod
 covers:
-  - desktop/macos/Desktop/Sources/Onboarding/OnboardingView.swift
   - desktop/macos/Desktop/Sources/Onboarding/OnboardingFlow.swift
+  - desktop/macos/Desktop/Sources/Onboarding/SecondBrain/SBOnboardingView.swift
   - desktop/macos/Desktop/Sources/DesktopAutomationBridge.swift
   - desktop/macos/Desktop/Sources/AppState/AppState+SystemActions.swift
   - desktop/macos/Desktop/Sources/OmiApp.swift
